% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets-forecast-censo.R
\name{ets_forecast_censo}
\alias{ets_forecast_censo}
\title{Exponential Smoothing Forecast}
\usage{
ets_forecast_censo(grupos, ano_FIRST_censo, ano_LAST_censo)
}
\arguments{
\item{grupos}{\code{list} contendo grupos de munícipios retornada pela função
\code{\link{split_mun_groups}}.}

\item{ano_FIRST_censo}{\code{vector} do tipo \code{numeric} contendo o ano
inicial da série, retornado pela função \code{\link{get_year_FIRST_censo}}.}

\item{ano_LAST_censo}{\code{vector} do tipo \code{numeric} contendo o ano
final da série, retornado pela função \code{\link{get_year_LAST_censo}}.}
}
\value{
\code{tibble} contendo o output dos diversos modelos aplicados nas
   etapas de ensino/escolas de um grupo específico de municípios.
}
\description{
Executa o forecast automático utilizando modelo de suavização
    exponencial.
}
\seealso{
Other forecast: \code{\link{tk_ts_transform_and_impute}}
}
\concept{forecast}
\keyword{internal}
