% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-and-clean-censo-full.R
\name{extract_and_clean_censo_full}
\alias{extract_and_clean_censo_full}
\title{Full Database Extraction and Cleaning}
\usage{
extract_and_clean_censo_full(tb_censo)
}
\arguments{
\item{tb_censo}{Objeto do tipo \code{tibble} do \code{dbplyr} que ainda vive
no SGBD, obtido por meio da função \code{\link{query_tb_censo}}}
}
\value{
\code{tibble} contendo a base completa do censo após tratamento e limpeza.
}
\description{
Trata a base do Censo e extrai do servidor para a máquina local
}
\examples{
\dontrun{

tb_censo_full <- extract_and_clean_censo_full(tb_censo)

}
}
\seealso{
Other query: 
\code{\link{query_tb_censo}()}
}
\concept{query}
