% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-year-FIRST-censo.R
\name{get_year_FIRST_censo}
\alias{get_year_FIRST_censo}
\title{Get the FIRST year of the time series}
\usage{
get_year_FIRST_censo(tb_censo_full)
}
\arguments{
\item{tb_censo_full}{\code{tibble} contendo a base completa do censo
após tratamento e limpeza por meio da função \code{\link{extract_and_clean_censo_full}}.}
}
\value{
\code{vector} do tipo \code{numeric} contendo o ano inicial da série.
}
\description{
Identifica e extrai o primeiro ano da série histórica do censo
}
\examples{
\dontrun{

ano_FIRST_censo <- get_year_FIRST_censo(tb_censo_full)

}
}
\seealso{
Other get_ano: \code{\link{get_year_LAST_censo}}
}
\concept{get_ano}
