% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-table-into-db.R
\name{insert_table_into_db}
\alias{insert_table_into_db}
\title{Insert Table into Data Base}
\usage{
insert_table_into_db(conn, tb_censo_final, schema, tab)
}
\arguments{
\item{conn}{String contendo a "string de conexão" retornada por
\code{\link{connect_censo_db}}.}

\item{tb_censo_final}{\code{tibble} contendo uma tabela final da projeção do
Censo Escolar. Pode ser a tabela técnica (TEC) ou de apresentação (APR).}

\item{schema}{String contendo o nome do schema do banco, ex:
\code{"INAPSISPRO_FNDE"}.}

\item{tab}{String contendo o nome da tabela que receberá o insert no banco.}
}
\value{
\code{TRUE} caso a gravação seja bem sucedida.
}
\description{
Grava a tabela informada no banco de dados
}
\examples{
\dontrun{

insert_table_into_db(conn, tb_censo_final, schema, tab)

}
}
\concept{writing}
