% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-final-table.R
\name{make_final_table}
\alias{make_final_table}
\title{Make Final Table}
\usage{
make_final_table(forecast_output, tb_censo, ano_FIRST_censo,
  ano_LAST_censo)
}
\arguments{
\item{forecast_output}{\code{list} de \code{tibble}s contendo o forecast output
de todos os modelos, a partir da execução de \code{\link{parallel_forecast}}.}

\item{tb_censo}{Objeto do tipo \code{tibble} do \code{dbplyr} que ainda vive
no SGBD, obtido por meio da função \code{\link{query_tb_censo}}}

\item{ano_FIRST_censo}{\code{vector} do tipo \code{numeric} contendo o ano
inicial da série, retornado pela função \code{\link{get_year_FIRST_censo}}.}

\item{ano_LAST_censo}{\code{vector} do tipo \code{numeric} contendo o ano
final da série, retornado pela função \code{\link{get_year_LAST_censo}}.}
}
\value{
Um \code{tibble} contendo os dados tratados.
}
\description{
Produz a tabela final contendo a projeção para gravação no SGBD
}
\examples{
\dontrun{

conn <- connect_censo_db(host, port, user, pass, svc)
tb_censo <- query_tb_censo(conn, schema, tab) # recuperando tab original para join
tb_censo_final_join <- make_final_table(forecast_output, tb_censo, ano_FIRST_censo, ano_LAST_censo)

}
}
\seealso{
Other final_wrangling: \code{\link{make_final_table_pres}}
}
\concept{final_wrangling}
