% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-final-table-pres.R
\name{make_final_table_pres}
\alias{make_final_table_pres}
\title{Make Final Table for Presentation}
\usage{
make_final_table_pres(tb_censo_final_join)
}
\arguments{
\item{tb_censo_final_join}{\code{tibble} contendo a tabela final com
especificações técnicas completas da projeção do Censo Escolar.}
}
\value{
Um \code{tibble} contendo os dados tratados para apresentação no SIMEC.
}
\description{
Produz a tabela final contendo a projeção para gravação no SGBD e
    apresentação no SIMEC
}
\examples{
\dontrun{

tb_censo_final_join_APR <- make_final_table_pres(tb_censo_final_join)

}
}
\seealso{
Other final_wrangling: \code{\link{make_final_table}}
}
\concept{final_wrangling}
