% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-tb-censo.R
\name{query_tb_censo}
\alias{query_tb_censo}
\title{Query the table containing the time series of all census data}
\usage{
query_tb_censo(conn, schema, tab)
}
\arguments{
\item{conn}{String contendo a "string de conexão" retornada por
\code{\link{connect_censo_db}}.}

\item{schema}{String contendo o nome do schema do banco, ex:
\code{"INAPSISPRO_FNDE"}.}

\item{tab}{String contendo o nome da tabela em que vive a base do censo, ex:
\code{"TB_CENSO_TS_QTD_ALUNOS"}.}
}
\value{
Objeto do tipo \code{tibble} do \code{dbplyr} que ainda vive no SGBD.
}
\description{
Executa uma query na base "INAPSISPRO_FNDE.TB_CENSO_TS_QTD_ALUNOS" no FNDE
}
\examples{
\dontrun{

schema = "INAPSISPRO_FNDE"
tab = "TB_CENSO_TS_QTD_ALUNOS"
tb_censo <- query_tb_censo(conn, schema, tab)

}
}
\seealso{
Other query: 
\code{\link{extract_and_clean_censo_full}()}
}
\concept{query}
