% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-mun-groups.R
\name{split_mun_groups}
\alias{split_mun_groups}
\title{Split Data in Groups}
\usage{
split_mun_groups(tb_censo_full, n_clusters)
}
\arguments{
\item{tb_censo_full}{\code{tibble} contendo a base completa do censo
após tratamento e limpeza por meio da função \code{\link{extract_and_clean_censo_full}}.}

\item{n_clusters}{String contendo a "string de conexão" retornada por
\code{\link{connect_censo_db}}.}
}
\value{
\code{tibble} contendo a base completa do censo após tratamento e limpeza.
}
\description{
Divide a base do Censo em grupos iguais ao número de cores como
    preparação para a execução paralelizada.
}
\examples{
\dontrun{

tb_censo_full <- extract_and_clean_censo_full(tb_censo)

}
}
\seealso{
Other clusters: 
\code{\link{detect_cores}()}
}
\concept{clusters}
